format prop changes
